Add OpenGL URLs to the extensions.
parse.php: try to find matches to GL or GLX extensions and see if a page exists
for this extension. If it exists, add a <link> child tag to <extension>.

index.php: modify extension name to add a link where the extension
appears.

config.example.php: add an "opengl_url" entry (in case there will be a
new "man" site somewhere else.