Modification date feature: backend
Parsing of GL3.txt now emits an XML with 'modified' attributes on supported
drivers and on the <mesa> element, which can be used in rendering to show when
a feature was implemented or otherwise changed.

This *does* mean lots of parsing of various versions of GL3.txt, but the parser
is quite fast anyway so caching isn't a high priority at the moment.