Change the presentation of the page.
I have gathered every information about the project into one section at
the bottom of the page: "About Mesamatrix". And then added sub-sections:
"How it works", "Source code", "How to help", "See also" and "Authors".

The date of the last parse of GL3.txt is written in the "How it works"
sub-section.

I kept a link to the full git log on top, under the commit list.