Parse the latest commit and define history from there
Before, commits were merged from the oldest to the most recent. This led
to some issues like some false-negatives on the supported drivers.

Now, the latest commit is parsed and then the other commits are compared
to this one (from the most recent to the oldest). If a change is
detected in the extension or its supported drivers, they are tagged with
the commit that induced the change.