Update grammar in ApiSubController.php
I saw this website linked in a news article and noticed some grammar issues. 

In the first sentence I changed "is" to "are" to fix the subject-verb agreement.

The way the 2nd sentence was written, specifically the phrase "by driver", could imply that the extensions were literally implemented by the driver and not for the driver. Due to AI/ML, it makes sense to write the sentence in a manner that implies the drivers themselves aren't programming their own extensions. I added the word "current" to make the sentence more specific.