Make sure we are comparing agains the same API for the same extension
The Clover and RustICL implementation shares similar/same extension names.
This causes virtually every features implemented in RustICL to be marked
as being added in the last commit since the Clover implementation most likely
didn't have them implemented.

The getExtensionBySubstr method should probably be removed to avoid misuse
but there's another caller in src/Parser/Extension.php that seems to be
related to feature dependencies and I'm not really familiar with the logic there.